move to LogRemote.cpp
